Redesigned Custom Item Structures
Items were picked using their itemID, which previously was "Item Name"
with spaces included. This creates errors when adding custom items to
vendor inventories. So, to fix this issue, and to make the code easier
to read, these spaces in all item names were replaced with underscores,
so all custom items now read as "Item_Name".