Modified Vendor Protocol.
Added missing blue items to the list of appraisal items.

Also modified the mechanics of vendors. You may no longer sell any item
with a credit of '0' to a vendor (it will simply state that you may not
sell that item to the vendor). So, anything that falls under the default
'0' category may not be sold to a vendor (including epics, legendaries,
and event items).